Refactor segment_anything.py
